+++ src/filter/filters/median/filter_median.cpp
@@ -18,7 +18,7 @@ void filter_median_c::apply(FILTER_FUNC_PARAMS) const
     VALIDATE_FILTER_INPUT
 
     #ifdef USE_OPENCV
-        const u8 kernelSize = this->parameter(PARAM_KERNEL_SIZE);
+        const double kernelSize = this->parameter(PARAM_KERNEL_SIZE);
 
         cv::Mat output = cv::Mat(r.h, r.w, CV_8UC4, pixels);
         cv::medianBlur(output, output, kernelSize);
